fix a crash when converting Color555 to Color

+++ Sources/Carbonizer/utilities/Color.swift
@@ -15,9 +15,9 @@ struct Color: Codable {
 	
 	init(_ rgb555: Color555) {
 		// * 8 doesn't map the entire range, it misses the highest 7 output values
-		red = rgb555.red * 255 / 31
-		green = rgb555.green * 255 / 31
-		blue = rgb555.blue * 255 / 31
+		red = UInt8(UInt(rgb555.red) * 255 / 31)
+		green = UInt8(UInt(rgb555.green) * 255 / 31)
+		blue = UInt8(UInt(rgb555.blue) * 255 / 31)
 	}
 	
 	enum ParseError: Error, CustomStringConvertible {
